+++ questionpy_server/worker/exception.py
@@ -10,13 +10,5 @@ class WorkerStartError(Exception):
     pass
 
 
-class WorkerMemoryLimitExceededError(Exception):
-    pass
-
-
 class WorkerCPUTimeLimitExceededError(Exception):
-    pass
-
-
-class WorkerUnknownError(Exception):
     pass
