+++ questionpy_server/worker/pool.py
@@ -10,10 +10,10 @@ from typing import Optional, Type
 from questionpy_common.constants import MiB
 from questionpy_common.environment import WorkerResourceLimits
 
+from questionpy_server.worker.runtime.package_location import PackageLocation
 from .exception import WorkerStartError
 from .worker import Worker
 from .worker.subprocess import SubprocessWorker
-from questionpy_server.worker.runtime.package_location import PackageLocation
 
 
 class WorkerPool:
